Refactor get_events endpoint
Requires 1 options params. Reason is once we
call BuildQuery.build we're passing in options (1 param) and we have
to verify what options contains at that BuildQuery level. Didn't see the benefit of
requiring params in the highest level when we don't expect them
to be present in the lower level